Changed some int to long.

+++ setsm_code.c
@@ -24,7 +24,7 @@
  * Algorithmica 2, 153-174.
  */
 
-#include "Function.h"
+#include "setsm_code.h"
 #include "math.h"
 #include <omp.h>
 #include <time.h>
@@ -6139,16 +6139,16 @@ bool VerticalLineLocus(NCCresult* nccresult, uint16 *MagImages_L,uint16 *MagImag
 #pragma omp parallel for schedule(guided)
 	for(long int iter_count = 0 ; iter_count < sub_imagesize_w*sub_imagesize_h ; iter_count++)
 	{
-		long int pts_row = (int)(floor(iter_count/sub_imagesize_w));
-		long int pts_col = iter_count % sub_imagesize_w;
-		int pt_index;
+		int pts_row = (int)(floor(iter_count/sub_imagesize_w));
+		int pts_col = iter_count % sub_imagesize_w;
+		long  pt_index;
 		float t_X, t_Y;
 		int t_col, t_row;
-		int pt_index_im;
+		long  pt_index_im;
 		
 		
-		t_X		= (float)((long)subBoundary[0] + pts_col*(long)im_resolution);
-		t_Y		= (float)((long)subBoundary[1] + pts_row*(long)im_resolution);
+		t_X		= subBoundary[0] + pts_col*im_resolution;
+		t_Y		= subBoundary[1] + pts_row*im_resolution;
 		
 		t_col	= (int)((t_X - subBoundary[0])/DEM_resolution);
 		t_row	= (int)((t_Y - subBoundary[1])/DEM_resolution);
@@ -6284,11 +6284,11 @@ bool VerticalLineLocus(NCCresult* nccresult, uint16 *MagImages_L,uint16 *MagImag
 	printf("height step %f\n",height_step);
 
 #pragma omp parallel for schedule(guided)
-	for(int iter_count = 0 ; iter_count < Size_Grid2D.height*Size_Grid2D.width ; iter_count++)
+	for(long  iter_count = 0 ; iter_count < Size_Grid2D.height*Size_Grid2D.width ; iter_count++)
 	{
 		int pts_row = (int)(floor(iter_count/Size_Grid2D.width));
 		int pts_col = iter_count % Size_Grid2D.width;
-		int pt_index = pts_row*Size_Grid2D.width + pts_col;
+		long  pt_index = pts_row*Size_Grid2D.width + pts_col;
 		
 		{
 			double pre_rho	= -1.0;
@@ -7210,7 +7210,7 @@ float VerticalLineLocus_seeddem(uint16 *MagImages_L,uint16 *MagImages_R,float DE
 	
 	int numofpts;
 	F2DPOINT *all_left_im_cd, *all_right_im_cd;
-	int sub_imagesize_w, sub_imagesize_h;
+	long sub_imagesize_w, sub_imagesize_h;
 	int pixel_buffer = 1000;
 	int GNCC_level	= 3;
 	bool check_ortho = false;
@@ -7237,14 +7237,14 @@ float VerticalLineLocus_seeddem(uint16 *MagImages_L,uint16 *MagImages_R,float DE
 	
 		
 #pragma omp parallel for schedule(guided)
-	for(int iter_count = 0 ; iter_count < sub_imagesize_w*sub_imagesize_h ; iter_count++)
+	for(long iter_count = 0 ; iter_count < sub_imagesize_w*sub_imagesize_h ; iter_count++)
 	{
 		int pts_row = (int)(floor(iter_count/sub_imagesize_w));
 		int pts_col = iter_count % sub_imagesize_w;
-		int pt_index;
+		long pt_index;
 		float t_X, t_Y;
 		int t_col, t_row;
-		int pt_index_im;
+		long pt_index_im;
 		
 		t_X		= subBoundary[0] + pts_col*im_resolution;
 		t_Y		= subBoundary[1] + pts_row*im_resolution;
@@ -7687,7 +7687,7 @@ bool VerticalLineLocus_blunder(float* nccresult, float* INCC, uint16 *MagImages_
 	
 	int numofpts;
 	F2DPOINT *all_left_im_cd, *all_right_im_cd;
-	int sub_imagesize_w, sub_imagesize_h;
+	long sub_imagesize_w, sub_imagesize_h;
 	int pixel_buffer = 1000;
 	int GNCC_level	= 3;
 	bool check_ortho = false;
@@ -7718,14 +7718,14 @@ bool VerticalLineLocus_blunder(float* nccresult, float* INCC, uint16 *MagImages_
 	
 	int pts_row, pts_col;
 #pragma omp parallel for schedule(guided)
-	for(int iter_count = 0 ; iter_count < sub_imagesize_w*sub_imagesize_h ; iter_count++)
+	for(long iter_count = 0 ; iter_count < sub_imagesize_w*sub_imagesize_h ; iter_count++)
 	{
 		int pts_row = (int)(floor(iter_count/sub_imagesize_w));
 		int pts_col = iter_count % sub_imagesize_w;
-		int pt_index;
+		long pt_index;
 		float t_X, t_Y;
 		int t_col, t_row;
-		int pt_index_im;
+		long pt_index_im;
 		
 		t_X		= subBoundary[0] + pts_col*im_resolution;
 		t_Y		= subBoundary[1] + pts_row*im_resolution;
